Rename @zero_iteration profile to the @loop
