+++ tests/phpunit/Metrics/SortableMutantExecutionResultsTest.php
@@ -38,7 +38,7 @@ namespace Infection\Tests\Metrics;
 use Infection\Metrics\SortableMutantExecutionResults;
 use Infection\Mutant\DetectionStatus;
 use Infection\Mutant\MutantExecutionResult;
-use Infection\Mutator\ZeroIteration\For_;
+use Infection\Mutator\Loop\For_;
 use Infection\Tests\Mutator\MutatorName;
 use PHPUnit\Framework\TestCase;
