+++ tests/phpunit/Mutator/Loop/While_Test.php
@@ -33,11 +33,11 @@
 
 declare(strict_types=1);
 
-namespace Infection\Tests\Mutator\ZeroIteration;
+namespace Infection\Tests\Mutator\Loop;
 
 use Infection\Tests\Mutator\BaseMutatorTestCase;
 
-final class Foreach_Test extends BaseMutatorTestCase
+final class While_Test extends BaseMutatorTestCase
 {
     /**
      * @dataProvider mutationsProvider
@@ -51,42 +51,51 @@ final class Foreach_Test extends BaseMutatorTestCase
 
     public function mutationsProvider(): iterable
     {
-        yield 'It mutates to new array in foreach' => [
+        yield 'It mutates expression part from variable to false' => [
             <<<'PHP'
 <?php
 
-$array = [1, 2];
-foreach ($array as $value) {
+$condition = true;
+
+while ($condition) {
 }
+
 PHP
             ,
             <<<'PHP'
 <?php
 
-$array = [1, 2];
-foreach (array() as $value) {
+$condition = true;
+while (false) {
 }
 PHP
-            ,
         ];
 
-        yield 'It does not change whether items were passed by reference' => [
+        yield 'It mutates expression part from boolean true to false' => [
             <<<'PHP'
 <?php
 
-$array = [1, 2];
-foreach ($array as $key => &$value) {
-    echo $value;
+while (true) {
 }
+
 PHP
             ,
             <<<'PHP'
 <?php
 
-$array = [1, 2];
-foreach (array() as $key => &$value) {
-    echo $value;
+while (false) {
 }
+PHP
+        ];
+
+        yield 'It does not mutate expression part in do-while loop to false' => [
+            <<<'PHP'
+<?php
+
+do {
+
+} while (true);
+
 PHP
         ];
     }
