+++ tests/phpunit/Mutator/MutatorResolverTest.php
@@ -43,9 +43,9 @@ use Infection\Mutator\Arithmetic\Plus;
 use Infection\Mutator\Boolean\IdenticalEqual;
 use Infection\Mutator\Boolean\NotIdenticalNotEqual;
 use Infection\Mutator\Boolean\TrueValue;
+use Infection\Mutator\Loop\For_;
 use Infection\Mutator\MutatorResolver;
 use Infection\Mutator\ProfileList;
-use Infection\Mutator\ZeroIteration\For_;
 use Infection\Tests\SingletonContainer;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
